+++ src/main/java/web/dao/UserDAOImpl.java
@@ -25,7 +25,7 @@ public class UserDAOImpl implements UserDAO {
     }
 
     @Override
-    public User show(int id) {
+    public User show(Long id) {
         User user = entityManager.find(User.class, id);
         return user;
     }
